More redux connect of AddContent into index.js
