+++ .eslintrc.json
@@ -16,7 +16,6 @@
     "jsx-a11y/anchor-is-valid": [ "error", {
       "aspects": [ "noHref" ]
     }],
-    "no-param-reassign": [2, { "props": false }],
-    "import/no-named-as-default": 0
+    "no-param-reassign": [2, { "props": false }]
   }
 }
