+++ src/components/05_pages/AddContent/index.js
@@ -1,0 +1,14 @@
+import { connect } from 'react-redux';
+
+import { requestContentTypes } from '../../../actions/application';
+
+import AddContent from './AddContent';
+
+const mapStateToProps = state => ({
+  contentTypes: state.application.contentTypes,
+});
+
+export default connect(
+  mapStateToProps,
+  { requestContentTypes },
+)(AddContent);
